requirements.txt
appdirs==1.4.0
docopt==0.6.2
packaging==16.8
progressbar2==3.12.0
pyparsing==2.1.10
python-utils==2.0.1
six==1.10.0
subprocess32==3.2.7
